Adicionar método de alterar representante

+++ src/main/java/com/br/zup/gerenciadordeguildas/services/RepresentanteService.java
@@ -1,6 +1,7 @@
 package com.br.zup.gerenciadordeguildas.services;
 
 import com.br.zup.gerenciadordeguildas.entities.Representante;
+import com.br.zup.gerenciadordeguildas.exceptions.RecursoNaoEncontradoException;
 import com.br.zup.gerenciadordeguildas.repositories.RepresentanteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,26 +14,29 @@ public class RepresentanteService {
     @Autowired
     private RepresentanteRepository representanteRepository;
 
-
     public Representante cadastrarRepresentante(Representante representante) {
         return representanteRepository.save(representante);
     }
 
-    public List<Representante> retornarTodosOsRepresentantes(){
-        List<Representante> representantes = (List<Representante>) representanteRepository.findAll();
-        return representantes;
+    public Iterable<Representante> retornarTodosOsRepresentantes(){
+       return representanteRepository.findAll();
     }
 
-    public Representante salvarRepresentante(Representante representante) {
-        try{
-            Representante obj = representanteRepository.save(representante);
+    public Representante atualizarRepresentante(Representante representante){
+        if(representanteRepository.existsById(representante.getId())){
+            Representante objRepresentante = representanteRepository.save(representante);
             return representante;
-        }catch (Exception error){
-            throw new RuntimeException("Representante já cadastrado!");
         }
+
+        throw new RecursoNaoEncontradoException("Representante", representante.getId());
     }
+
     public void deletarRepresentante(Integer id) {
-        representanteRepository.deleteById(id);
+        if(representanteRepository.existsById(id)){
+            representanteRepository.deleteById(id);
+        }
+
+        throw new RecursoNaoEncontradoException("Representante", id);
     }
 
 }
